Move ByteChunkSyntax to the org.http4s.syntax package
To be consistent with other syntax enrichments
