+++ blaze-core/src/main/scala/org/http4s/blazecore/util/Http2Writer.scala
@@ -7,7 +7,7 @@ import fs2._
 import org.http4s.blaze.http.Headers
 import org.http4s.blaze.http.http20.NodeMsg._
 import org.http4s.blaze.pipeline.TailStage
-import org.http4s.util.chunk._
+import org.http4s.syntax.byteChunk._
 import scala.concurrent._
 
 private[http4s] class Http2Writer[F[_]](
