+++ blaze-core/src/main/scala/org/http4s/blazecore/util/IdentityWriter.scala
@@ -8,7 +8,7 @@ import fs2._
 import java.nio.ByteBuffer
 import org.http4s.blaze.pipeline.TailStage
 import org.http4s.util.StringWriter
-import org.http4s.util.chunk._
+import org.http4s.syntax.byteChunk._
 import org.log4s.getLogger
 import scala.concurrent.{ExecutionContext, Future}
