+++ core/src/main/scala/org/http4s/syntax/AllSyntax.scala
@@ -4,6 +4,7 @@ package syntax
 trait AllSyntax
     extends AnyRef
     with AsyncSyntax
+    with ByteChunkSyntax
     with EffectResponseSyntax
     with EffectRequestSyntax
     with KleisliSyntax
