+++ core/src/main/scala/org/http4s/syntax/package.scala
@@ -3,6 +3,7 @@ package org.http4s
 package object syntax {
   object all extends AllSyntax
   object async extends AsyncSyntax
+  object byteChunk extends ByteChunkSyntax
   @deprecated("Use map or flatMap on the request instead", "0.18.0-M2")
   object effectRequest extends EffectRequestSyntax
   @deprecated("Use map or flatMap on the request instead", "0.18.0-M2")
